+++ sdk/storage/azure-storage-blob-cryptography/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.33.0-beta.2 (Unreleased)
+## 12.34.0-beta.2 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,16 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 12.32.4 (2026-05-05)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-storage-blob` from `12.33.3` to version `12.33.4`.
+- Upgraded `azure-core-http-netty` from `1.16.3` to version `1.16.4`.
+- Upgraded `azure-core` from `1.57.1` to version `1.58.0`.
 
 ## 12.32.3 (2026-03-30)